refactor(graph-page): refactor graph components structure for independ styles and logic changes
