+++ src/pages/Graph/App.scss
@@ -1,18 +1,3 @@
-$myColor: rgb(17, 0, 255);
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  color: $myColor;
-  text-align: center;
-}
-
-h1 {
-  margin-top: 3rem;
-}
 
 .App {
   background-color: #282c34;
