+++ src/pages/Graph/views/MindGraph/containers/NavBar/index.scss
@@ -1,0 +1,17 @@
+@import '../../../../../../styles.config.scss';
+
+.nav-bar {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+
+    h1 {
+        margin-top: 1rem;
+        color: $primary-color;
+        text-align: left;
+        font-size: 1.5rem;
+        padding-left: 2rem;
+      }
+      
+}
