+++ src/pages/Graph/views/MindGraph/containers/NavBar/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import './index.scss'
+
+export const NavBar: React.FC = () => (
+    <div className="nav-bar">
+        <h1>Mind Graph</h1>
+    </div>
+)
